fix: Move CustomTooltip into Chart folder
